testing solution

+++ 0x14-file_io/0-read_textfile.c
@@ -10,7 +10,7 @@
 ssize_t read_textfile(const char *filename, size_t letters)
 {
 	int fd;
-	ssize_t length;
+        int length;
 	char *buffer;
 
 	if (filename == NULL)
